+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/mimex/ContentType.java
@@ -19,6 +19,7 @@ public enum ContentType {
     UNIT_COMMAND,
     LACCESS,
     LVAR,
+    WEATHER,
     ;
 
     public final boolean hasLookup;